+++ Assets/GameAssets/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8074988b85b4ce6438fc9a3fe1e6182f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
