+++ Assets/GameAssets/Sprites/CannonLevel/piratelongrope.png.meta
