+++ Assets/GameAssets/Sprites/Common.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5ef749b69f0e55c419d3bc52ffda2936
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
